requirements.txt
bs4
indic-transliteration
sentence_transformers
faiss-cpu
numpy
openai
dotenv
flask
flask-cors